マイグレーションファイル修正・rollback時のエラー対応

+++ database/migrations/2023_04_10_030329_add_percent_to_items_table.php
@@ -14,7 +14,7 @@ class AddPercentToItemsTable extends Migration
     public function up()
     {
         Schema::table('items', function (Blueprint $table) {
-            $table->unsignedInteger('percent');
+            $table->unsignedInteger('percent')->default(30);
         });
     }
 
@@ -25,8 +25,10 @@ class AddPercentToItemsTable extends Migration
      */
     public function down()
     {
-        Schema::table('items', function (Blueprint $table) {
-            $table->dropColumn('percent');
-        });
+        if (Schema::hasTable('items')) {
+            Schema::table('items', function (Blueprint $table) {
+                $table->dropColumn('percent');
+            });
+        }
     }
 }
